make vk returned by zkprogram consistent with smart contract
